requirements.txt
absl-py==2.0.0
adal==1.2.7
addict==2.4.0
alembic==1.12.0
ansi2html==1.8.0
anyio==4.0.0
appdirs==1.4.4
applicationinsights==0.11.10
argcomplete==2.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
av==10.0.0
azure-ai-ml==1.10.1
azure-common==1.1.28
azure-core==1.29.4
azure-graphrbac==0.61.1
azure-identity==1.14.0
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.3
azure-mgmt-network==21.0.1
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
-e git+ssh://git@ssh.dev.azure.com/v3/devsdb/CRD-NT_EPFL-Handbook/azure-ml-utils@ee3f2d4de6a523b5986a472548c8c3bf68254701#egg=azure_ml_utils
azure-storage-blob==12.13.0
azure-storage-file-datalake==12.13.1
azure-storage-file-share==12.14.1
azureml-automl-core==1.53.0
azureml-dataprep==4.12.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.19.3
azureml-dataset-runtime==1.53.0
azureml-mlflow==1.53.0
azureml-pipeline==1.53.0
azureml-pipeline-core==1.53.0
azureml-pipeline-steps==1.53.0
azureml-telemetry==1.53.0
azureml-train-automl-client==1.53.0
azureml-train-core==1.53.0
azureml-train-restclients-hyperdrive==1.53.0
Babel==2.12.1
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
bpy==3.6.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.5
colorama==0.4.6
colorlog==6.7.0
comet-ml==3.33.10
comm==0.1.4
ConfigArgParse==1.7
configobj==5.0.8
contextlib2==21.6.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
Cython==3.0.2
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
distro==1.8.0
docker==6.1.3
docker-pycreds==0.4.0
docstring-parser==0.15
dotnetcore2==3.1.23
dulwich==0.21.6
embreex==2.17.7.post3
entrypoints==0.4
everett==3.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
fire==0.5.0
Flask==2.2.5
fonttools==4.42.1
fqdn==1.5.1
frozendict==2.3.8
fusepy==3.0.1
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.36
google-api-core==2.11.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
grpcio==1.58.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaxtyping==0.2.21
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
knack==0.10.1
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.41.0
lpips==0.1.4
lxml==4.9.3
Mako==1.2.4
mapbox-earcut==1.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapy==1.1.9
mistune==3.0.1
mlflow==2.2.2
mlflow-skinny==2.7.1
mpmath==1.3.0
msal==1.24.0
msal-extensions==1.0.0
msgpack==1.0.5
msgpack-numpy==0.4.8
msrest==0.7.1
msrestazure==0.6.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.7.0
ndg-httpsclient==0.5.1
nerfacc==0.5.2
-e git+https://github.com/nerfstudio-project/nerfstudio.git@e0d8f0426752dc0f97bd332fd6b84e95e8377210#egg=nerfstudio
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1
nodeenv==1.8.0
notebook==7.0.4
notebook_shim==0.2.3
numba==0.58.0
numpy==1.23.5
nuscenes-devkit==1.1.9
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
open3d==0.17.0
opencensus==0.11.3
opencensus-context==0.1.3
opencensus-ext-azure==1.1.10
opencv-python==4.6.0.66
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkginfo==1.9.6
platformdirs==3.10.0
plotly==5.17.0
portalocker==2.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycollada==0.7.2
pycparser==2.21
pydash==5.1.2
pyequilib==0.5.7
Pygments==2.16.1
PyJWT==2.8.0
pyliblzfse==0.4.1
pymeshlab==2022.2.post4
PyNaCl==1.5.0
pyngrok==6.1.2
pyntcloud==0.3.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.8.2
python-engineio==4.7.1
python-json-logger==2.0.7
python-socketio==5.9.0
pytransform3d==3.4.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
querystring-parser==1.2.4
rawpy==0.18.1
-e git+ssh://git@github.com/SchindlerEPFL/rebel-nerf.git@26e1d30c544fe50fa9dec5c8ed50a03b19a9f0cc#egg=rebel_nerf
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
Rtree==1.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
SecretStorage==3.3.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.31.0
setproctitle==1.3.2
shap==0.42.1
shapely==2.0.1
shtab==1.6.4
simple-websocket==0.10.1
simplejson==3.19.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
splines==0.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.2
strictyaml==1.7.3
svg.path==6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.18
timm==0.6.7
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==1.1.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trimesh==3.23.5
triton==2.0.0
typeguard==4.1.5
typing_extensions==4.8.0
tyro==0.5.7
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
viser==0.1.3
wandb==0.15.10
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.3.3
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wrapt==1.15.0
wsproto==1.2.0
wurlitzer==3.0.3
xatlas==0.0.8
xxhash==3.3.0
yourdfpy==0.0.53
zipp==3.17.0
zstandard==0.21.0
